+++ routes/api/decks.js
@@ -6,16 +6,12 @@ const ensureLoggedIn = require('../../config/ensureLoggedIn');
 //POST /api/decks
 router.post('/', ensureLoggedIn, decksCtrl.create);
 //GET /api/decks
-router.get('/', ensureLoggedIn, decksCtrl.getAllDecks); 
-//POST /api/decks/cards
-router.post('/cards', ensureLoggedIn, decksCtrl.createCard);
-//GET /api/decks/id
-router.get('/:id', ensureLoggedIn, decksCtrl.getCards);
+router.get('/', ensureLoggedIn, decksCtrl.getMyDecks); 
 //DELETE /api/decks/id
 router.delete('/:id', ensureLoggedIn, decksCtrl.deleteDeck);
 //PUT /api/decks
 router.put('/:id', ensureLoggedIn, decksCtrl.editDeckName);
-//DELETE /api/decks/card/id
-router.delete('/card/:id', ensureLoggedIn, decksCtrl.deleteCard)
+//GET /api/decks/all-decks
+router.get('/all-decks', ensureLoggedIn, decksCtrl.getAllDecks);
 
 module.exports = router;
