+++ src/pages/DeckDetailsPage/DeckDetailsPage.jsx
@@ -1,6 +1,6 @@
 import { useParams } from "react-router-dom";
 import { useEffect } from 'react';
-import { getCards } from '../../utilities/decks-api';
+import { getCards } from '../../utilities/cards-api';
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import Card from '../../components/Card/Card';
